added

+++ app/Http/Controllers/dashboardcontroller.php
@@ -4,29 +4,41 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\smscontroller;
-use App\Models\{AgentUser,User,Account,Fleet,Category,Route,Calendarial,AgentRoute};
+use App\Models\{AgentUser,User,Account,Fleet,Category,Route,Calendarial,AgentRoute,Booking,Parcel};
 use DB;
 use Hash;
 use Session;
 use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 use Illuminate\Support\{Str};
+use Auth;
 
 class dashboardcontroller extends Controller
 {
     public function __construct() {
         $this->middleware('auth');
     }
+    public function check_if_admin() {
+        if(Auth::check() && Auth::user()->hasRole('admin')) {
+            return true;
+        }
+        return false;
+    }
     public function index() {
         $user = auth()->user();
         $routes = Route::where([['user_id', $user->id]])->with('agent')->get();
-        return view('dashboard.index', compact('routes'));
+        $bookings = Booking::where('user_id', auth()->user()->id)->get()->count();
+        $parcels = Parcel::where('user_id', auth()->user()->id)->get()->count();
+        return view('dashboard.index', compact('routes', 'bookings', 'parcels'));
     }
     public function agents() {
+        if($this->check_if_admin() == false) return redirect()->back();
         $user = auth()->user();
         $agents = AgentUser::where('company_id', $user->id)->with('user')->get();
         return view('dashboard.agents', compact('agents'));
     }
     public function add_agent(Request $request) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $this->validate($request, [
             'email' => 'required',
             'id_no' => 'required',
@@ -39,7 +51,7 @@ class dashboardcontroller extends Controller
         $user = auth()->user();
         $handle = strtolower(Str::random(7));
         $agent_unique = Str::random(7);
-        $password = strtolower(substr($request->first_name, 0, 2).mt_rand(1000,9999));
+        $password = strtolower(substr($request->fname, 0, 2).mt_rand(1000,9999));
         DB::transaction(function() use($request,$password,$user,$handle,$agent_unique) {
             $user_data = [
                 'fname' => $request->fname,
@@ -83,7 +95,8 @@ class dashboardcontroller extends Controller
                 'company_id' => $user->id,
                 'user_id' => $new_user->id,
                 'agent_unique' => strtolower($agent_unique),
-                'office_id' => $request->office_id
+                'office_id' => $request->office_id,
+                'pass_code' => $password
             ]);
         });
         $sms = new smscontroller;
@@ -95,11 +108,13 @@ class dashboardcontroller extends Controller
     }
     public function agent_lock($id) {}
     public function agent_unlock($id) {}
-    public function fleets() {
+    public function add_fleets() {
+        if($this->check_if_admin() == false) return redirect()->back();
         $fleets = Fleet::orderBy('id', 'desc')->where('user_id', auth()->user()->id)->get();
         return view('dashboard.fleets', compact('fleets'));
     }
     public function add_fleet(Request $request) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $this->validate($request, [
             'fleet_id' => 'required|unique:fleets',
             'driver_name' => 'required',
@@ -117,18 +132,21 @@ class dashboardcontroller extends Controller
         return redirect()->back();
     }
     public function unsuspend_fleet($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         Fleet::where('id', base64_decode($id))->update(['suspend' => false]);
         Session::flash('success', 'Fleet unsuspended successfully.');
         return redirect()->back();
     }
     public function suspend_fleet($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         Fleet::where('id', base64_decode($id))->update(['suspend' => true]);
         Session::flash('success', 'Fleet suspended successfully.');
         return redirect()->back();
     }
     public function edit_fleet(Request $request, $id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $this->validate($request, [
-            'fleet_id' => 'required|unique:fleets',
+            'fleet_id' => 'required',
             'driver_name' => 'required',
             'driver_contact' => 'required|digits:10'
         ]);
@@ -143,12 +161,14 @@ class dashboardcontroller extends Controller
         return redirect()->back();
     }
     public function delete_fleet($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $fl = Fleet::find(base64_decode($id));
         $fl->delete();
         Session::flash('success', 'Fleet deleted successfully.');
         return redirect()->back();
     }
     public function calendarial() {
+        if($this->check_if_admin() == false) return redirect()->back();
         $category = Category::get();
         $route = Route::where([['user_id', auth()->user()->id]])->get();
         $calendar = Category::with('calendarial')->get();
@@ -159,6 +179,7 @@ class dashboardcontroller extends Controller
         return view('dashboard.calendarial', compact('category', 'route', 'data'));
     }
     public function add_peak(Request $request) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $route = Route::find($request->fleet_id);
         $exists = Calendarial::where([
             ['user_id', auth()->user()->id],
@@ -177,17 +198,34 @@ class dashboardcontroller extends Controller
             'amount' => $request->amount,
             'user_id' => auth()->user()->id,
             'off_peak' => $route->amount,
-            'name' => $name
+            'name' => $name,
+            'lock' => $request->lock
         ];
         Calendarial::create($data);
         Session::flash('success', 'Peak added successfully.');
         return redirect()->back();
     }
+    public function edit_peak(Request $request, $id) {
+        if($this->check_if_admin() == false) return redirect()->back();
+        Calendarial::where('id', base64_decode($id))->update([
+            'date' => $request->date,
+            'amount' => $request->amount
+        ]);
+        Session::flash('success', 'Edited Successfully.');
+        return redirect()->back();
+    }
+    public function delete_peak($id) {
+        Calendarial::find(base64_decode($id))->delete();
+        Session::flash('error', 'Deleted Successfully.');
+        return redirect()->back();
+    }
     public function add_route() {
+        if($this->check_if_admin() == false) return redirect()->back();
         $agents = AgentUser::where('company_id', auth()->user()->id)->with('user')->get();
         return view('dashboard.routes.add_route', compact('agents'));
     }
     public function create_route(Request $request) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $this->validate($request, [
             'group' => 'required|min:3',
             'amount' => 'required',
@@ -200,14 +238,14 @@ class dashboardcontroller extends Controller
         ]);
         $user = auth()->user();
         $routes = Route::where('user_id', $user->id)->get();
+        $unique = 'FLID'.mt_rand(1000,9999);        
         foreach($routes as $route) {
             if($route->seaters == $request->seaters && $route->departure == $request->departure && $route->destination == $request->destination) {
                 Session::flash('error', 'Oops, fleet with similar details exists.');
                 return redirect()->back();
             }
         }
-        DB::transaction(function() use($request,$user) {
-            $unique = 'FLID'.mt_rand(1000,9999);
+        DB::transaction(function() use($request,$user,$unique) {            
             $route_data = [
                 'user_id' => $user->id,
                 'group' => $request->group,
@@ -241,16 +279,19 @@ class dashboardcontroller extends Controller
         return redirect()->route('dashboard.routes');
     }
     public function routes() {
-        $routes = Route::where('user_id', auth()->user()->id)->get();
+        if($this->check_if_admin() == false) return redirect()->back();
+        $routes = Route::where('user_id', auth()->user()->id)->with('agent')->get();
         return view('dashboard.routes.routes', compact('routes'));
     }
     public function edit_route($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $route = Route::find(base64_decode($id));
         $agents = AgentUser::where('company_id', auth()->user()->id)->with('user')->get();
         $agent = AgentRoute::where('route_id', base64_decode($id))->with('user')->first();
         return view('dashboard.routes.edit_route', compact('route', 'agents', 'agent'));
     }
     public function edit_route_post(Request $request, $id) {
+        if($this->check_if_admin() == false) return redirect()->back();
         $this->validate($request, [
             'group' => 'required|min:3',
             'amount' => 'required',
@@ -304,4 +345,91 @@ class dashboardcontroller extends Controller
         //delete also relation with user
     }
     public function top_up_agent(Request $request) {}
+    public function suspend_route($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
+        $route = Route::find(base64_decode($id));
+        $route->suspend = true;
+        $route->save();
+        Session::flash('info', 'Route locked.');
+        return redirect()->back();
+    }
+    public function unsuspend_route($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
+        $route = Route::find(base64_decode($id));
+        $route->suspend = false;
+        $route->save();
+        Session::flash('info', 'Route unlocked.');
+        return redirect()->back();
+    }
+    public function admin_suspend_route($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
+        $route = Route::find(base64_decode($id));
+        $route->admin_suspend = true;
+        $route->save();
+        Session::flash('info', 'Route locked.');
+        return redirect()->back();
+    }
+    public function admin_unsuspend_route($id) {
+        if($this->check_if_admin() == false) return redirect()->back();
+        $route = Route::find(base64_decode($id));
+        $route->admin_suspend = false;
+        $route->save();
+        Session::flash('info', 'Route unlocked.');
+        return redirect()->back();
+    }
+    public function booking_office(Request $request) {
+        $route = Route::where([['id', $request->route_id], ['admin_suspend', false]])->get();
+        return json_encode($route);
+    }
+    public function view_ticket_7($id) {
+        $route = Route::find($id);
+        return view('tickets.7', compact('route'));
+    }
+    public function view_ticket_10($id) {
+        return view('tickets.10');
+    }
+    public function view_ticket_11($id) {
+        return view('tickets.11');
+    }
+    public function view_ticket_14($id) {
+        return view('tickets.14');
+    }
+    public function view_ticket_16($id) {
+        return view('tickets.16');
+    }
+    public function bookings() {
+        $bookings = Booking::where('user_id', auth()->user()->id)->get();
+        return view('dashboard.bookings', compact('bookings'));
+    }
+    public function fleets() {
+        return view('dashboard.fleet_dispatches');
+    }
+    public function parcels() {
+        return view('dashboard.parcel_dispatches');
+    }
+    public function wallet() {
+        return view('dashboard.wallet');
+    }
+    public function edit_account() {
+        $user = auth()->user();
+        return view('dashboard.edit_account', compact('user'));
+    }
+    public function update_account(Request $request, $id) {
+        $this->validate($request, [
+            'mobile' => 'required|digits:10',
+            'c_mobile' => 'required|digits:10',
+            'fname' => 'required',
+            'lname' => 'required',
+            'password' => 'required'
+        ]);
+        User::where('id', base64_decode($id))->update([
+            'fname' => $request->fname,
+            'lname' => $request->lname,
+            'mobile' => $request->mobile,
+            'c_mobile' => $request->c_mobile,
+            'password' => Hash::make($request->password)
+        ]);
+        Session::flash('info', 'Account updated successfully.');
+        return redirect()->route('dashboard.index');
+    }
 }
